Move payment checker scheduler to module
