--- modules/ppcp-saved-payment-checked/extensions.php
@@ -1,14 +1,0 @@
-<?php
-/**
- * The SavedPaymentChecker module extensions.
- *
- * @package WooCommerce\PayPalCommerce\SavedPaymentChecker
- */
-
-declare(strict_types=1);
-
-namespace WooCommerce\PayPalCommerce\SavedPaymentChecker;
-
-use WooCommerce\PayPalCommerce\Vendor\Psr\Container\ContainerInterface;
-
-return array();
