+++ modules/ppcp-saved-payment-checker/extensions.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * The SavedPaymentChecker module services.
+ * The SavedPaymentChecker module extensions.
  *
  * @package WooCommerce\PayPalCommerce\SavedPaymentChecker
  */
@@ -8,7 +8,5 @@
 declare(strict_types=1);
 
 namespace WooCommerce\PayPalCommerce\SavedPaymentChecker;
-
-use WooCommerce\PayPalCommerce\Vendor\Psr\Container\ContainerInterface;
 
 return array();
