+++ modules/ppcp-wc-gateway/src/Gateway/PayPalGateway.php
@@ -556,9 +556,7 @@ class PayPalGateway extends \WC_Payment_Gateway {
 				);
 			}
 
-			if ( $this->subscription_helper->has_subscription( $order_id ) ) {
-				$this->schedule_saved_payment_check( $order_id, $wc_order->get_customer_id() );
-			}
+			do_action( 'woocommerce_paypal_payments_before_handle_payment_success', $wc_order );
 
 			return $this->handle_payment_success( $wc_order );
 		} catch ( PayPalApiException $error ) {
